requirements.txt
# Requirements for Round 1B: Persona-Driven Document Intelligence
# All packages are CPU-only and lightweight (<1GB total)

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pdfminer.six==20221105

# NLP and Text Processing (CPU-friendly)
nltk==3.8.1
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Mathematical and data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Standard libraries (already included in Python)
# json, os, time, re, pathlib, datetime, typing, warnings, collections
